Make enumeration EColumnType in class MyMoney Report type safe by converting to a class
Summary:
The recent implementation of report related enumerations is not type safe,
because enums are implemented as unsigned int, which allows to assign any
enum constant to any enum variable and open the door for wrong assignments.
Type safe enumaration reduces the number possible errors in this area
on refactoring or bug fixing.

This patch introduces a class below MyMoneyReport to separate the
enumerations from each other. Namespaces inside classes are not allowed
and Q_ENUM(s) or QFlags are also not usable, because the related enums are
required to be defined somewhere outside, which does not solve the issue.

CCBUG:391091

Test Plan: compiled on linux

Reviewers: #kmymoney

Differential Revision: https://phabricator.kde.org/D11043

Signed-off-by: Ralf Habacker <ralf.habacker@freenet.de>