Add support for checking a selected online quote in related settings dialog
As additional edit features the online quote settings now has a button to
show the related source in a web browser and a button to be able to delete
obsolate online quotes.

To support this feature, the WebPriceQuote class has been extended by detailed
error results and the output of status and error messages has been corrected.

BUG:388449
FIXED-IN:4.8.3

Differential Revision: https://phabricator.kde.org/D13558