Fix ignoring category filter in transaction by 'month reports'
Summary:
The issue was caused because there was no implementation of
MyMoneyDatabaseMgr::transactionList() returning a modified list
of splits for QueryTable::constructTransactionTable().

There is a bug in MyMoneyTransactionFilter in case a report uses
setInvestmentsOnly(true) not returning the requested splits.
To work arround this issue, the original splits are used in
QueryTable::constructTransactionTable() in this specific case.

BUG:395327
FIXED-IN:4.8.3

Test Plan: compiled and verified with test case from related bug and querytabletest

Reviewers: tbaumgart

Differential Revision: https://phabricator.kde.org/D13803