Add support for checking reports defined in a kmymoney file
This test loads available kmymoney test files, iterates through the
embedded reports and compares the generated content with a reference
file and reports differences.

The reference file in XML format contains the internal rows of the class
ListTable, which forms the basis for ListTable and QueryTable-related
reports.

Additional test files are created with kmymoney (e.g. based on test cases)
and stored in kmymoney/reports/testfiles. Then call'reporttest', which
attempts to find appropriate reference files for the embedded reports to
compare. If no reference files are available, a corresponding template
for a reference file is generated by reporttest in the specified kmymoney
source directory.

If a reference file is found and the check fails, reporttest creates an
additional XML file with the addition '-current.xml'. This file can then
be easily compared to the reference file containing the addition
'-reference.xml' for troubleshooting purposes.

Possible extensions:
- Add support for Pivotable based reports
- Improved numerical comparison with different denominators
- Ignore individual rows and columns
- Creating the reference file using the report export function