Add support for a predefined fiscal year as report date range
When reports are created, they are often defined with the current (fiscal)
year. If the KMyMoney file is opened in subsequent years, such customized
reports are no longer available and must be redefined individually for the
corresponding year.

To surround the problem, a new "Predefined Fiscal Year" date range is
added, the range of which can be specified on the File Information page
and used in reports, specifying stable range boundaries over years.