Add hint for marker transaction
Marker transactions contain only a split and a
null value and are used for identification
according to information provided by Thomas
Baumgart.

The problem with this is that there seems to be
no way to create a new transaction of this type
via the account book or to save a text change.