Add support for opening URLs from identifiers in transaction notes
This feature adds support for payee based links in transactions.

This is done by adding a template for a link e.g.
'https://www.heise.de/select/ct/archiv/%1' and a regular expression for an
identifier e.g. '(20\d\d/\d+)\b' in the new "Link Matching" tab of the
associated payee to catch the term '20xx/yy'.

In transactions associated with this recipient, the corresponding term,
e.g. '2022/15' can then be added to get an underlined link in the view.

The link is opened after selecting the corresponding transaction by the
new entry "Open URL" in the context menu or by using a keyboard shortcut.

Enable open url action, when one transaction is selected

CCBUG:437481