Add enabled components to component page of the application info dialog
On the components page, the components used when building are now
displayed next to KF5 and Qt5.
For the Alkimia library, the version used for building and the version
used at runtime are queried and displayed via include header/API. For the
remaining components, only the version used for building can be displayed,
as the corresponding API functions are missing.

BUG:491978
FIXED-IN:5.2