cmake: Add support to generate the doc as normal html page
A cmake macro named kmm_kdoctools_install() was added to generate
the HTML documentation.
To obtain complete documentation, additional files must be copied
to the same location and according to a specific logic to avoid
missing images, which is not supported by kdoctools_install().

This support is primarily intended for developers to avoid
the more complex detour via meinproc.