Add "Rate" column to reports
This commit introduces a new report column called “Rate,”
which is used in place of the “Price” column in reports
that include currency conversions. When the “Rate” column
is visible, discrepancies caused by unexpected exchange
rates are self-explanatory.

The “Rate” column is enabled by default if, for a report:
- the “Convert to base currency” option is selected,
- the report contains accounts with more than one currency, and
- exchange rates are available for the currency pairs included

The “Price” column is enabled by default if, for a report:
- the “Convert to base currency” option is selected,
- investment or loan accounts are included,
- stock prices are present in investment or loan reports.

In all other cases, these columns can be added or removed.

BUG:506692
BUG:507881
BUG:507935
BUG:507938
FIXED-IN:5.2.3